+++ data-pojo/src/main/java/com/yexca/dto/PersonalUserUpdateDTO.java
@@ -14,11 +14,11 @@ public class PersonalUserUpdateDTO implements Serializable {
 
     private String email;
 
-    private Integer phone;
+    private Long phone;
 
     private Integer gender;
 
-    private Integer countryId;
+    private Long countryId;
 
     private String nin;
