+++ data-pojo/src/main/java/com/yexca/vo/EmployeeUpdateVO.java
@@ -18,11 +18,11 @@ public class EmployeeUpdateVO implements Serializable {
 
     private String email;
 
-    private Integer phone;
+    private Long phone;
 
     private Integer gender;
 
-    private Integer countryId;
+    private Long countryId;
 
     private String nin;
